Refactor to use BigDecimal for currency calculations
Replaced Float with BigDecimal in all currency-related logic to improve precision. Updated relevant methods, classes, and APIs to handle the new data type, ensuring compatibility and correctness. Added periodic cache updates and removed deprecated expiration logic from the CurrencyRateCache.

+++ backend/src/main/java/com/mikhaile/nostobackend/ConvertResponse.java
@@ -1,9 +1,11 @@
 package com.mikhaile.nostobackend;
 
+import java.math.BigDecimal;
+
 record ConvertResponse(
     String baseCurrency,
     String quoteCurrency,
-    Float baseAmount,
-    Float quoteAmount
+    BigDecimal baseAmount,
+    BigDecimal quoteAmount
 ) {
 }
